+++ x/pylons/client/cli/loud_game_test.go
@@ -86,7 +86,6 @@ func createLOUDCookbook(t *testing.T, simInfo *loudBasicSim) {
 		"Pylons Inc",
 		"v0.0.1",
 		"alex@shmeeload.xyz",
-		"{\"denom\": \"pylons\", \"amount\": \"1\"}",
 		"true",
 	}
 
@@ -162,6 +161,7 @@ func createCharacterRecipe(t *testing.T, simInfo *loudBasicSim) {
 		string(entries),
 		string(itemOutputs),
 		"0",
+		"{\"denom\": \"pylons\", \"amount\": \"1\"}",
 		"true",
 		"extraInfo",
 	}
@@ -245,6 +245,7 @@ func getLOUDCoin(t *testing.T, simInfo *loudBasicSim) {
 		string(entries),
 		string(outputs),
 		"0",
+		"{\"denom\": \"pylons\", \"amount\": \"1\"}",
 		"true",
 		"extraInfo",
 	}
@@ -344,6 +345,7 @@ func createBuyCopperSwordRecipe(t *testing.T, simInfo *loudBasicSim) {
 		string(entries),
 		string(itemOutputs),
 		"0",
+		"{\"denom\": \"pylons\", \"amount\": \"1\"}",
 		"true",
 		"extraInfo",
 	}
@@ -596,6 +598,7 @@ func fightWolfWithSword(t *testing.T, simInfo *loudBasicSim) {
 		string(entries),
 		string(outputs),
 		"0",
+		"{\"denom\": \"pylons\", \"amount\": \"1\"}",
 		"true",
 		"extraInfo",
 	}
@@ -607,7 +610,6 @@ func fightWolfWithSword(t *testing.T, simInfo *loudBasicSim) {
 	_, err = clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdCreateRecipe(), args)
 	require.NoError(t, err)
 
-	var character types.Item
 	// Farm this wolf fight
 	for i := 0; i < 30; i++ {
 		// get sword and character IDs
@@ -652,7 +654,6 @@ func fightWolfWithSword(t *testing.T, simInfo *loudBasicSim) {
 		charOut, err := clitestutil.ExecTestCLICmd(simInfo.ctx, cli.CmdShowItem(), args)
 		require.NoError(t, err)
 		require.NoError(t, simInfo.ctx.JSONCodec.UnmarshalJSON(charOut.Bytes(), &itemResp))
-		character = itemResp.Item
 		if itemResp.Item.Owner != simInfo.net.Validators[0].Address.String() {
 			// PLAYER DIED
 			createCharacter(t, simInfo)
@@ -670,6 +671,4 @@ func fightWolfWithSword(t *testing.T, simInfo *loudBasicSim) {
 			continue
 		}
 	}
-
-	fmt.Println(character)
 }