+++ x/pylons/keeper/msg_server_recipe_test.go
@@ -30,7 +30,6 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerCreate() {
 			Developer:    "",
 			Version:      "v0.0.1",
 			SupportEmail: "test@email.com",
-			CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 			Enabled:      false,
 		}
 		_, err := srv.CreateCookbook(wctx, cookbook)
@@ -47,6 +46,7 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerCreate() {
 			Entries:       types.EntriesList{},
 			Outputs:       nil,
 			BlockInterval: 0,
+			CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 			Enabled:       false,
 			ExtraInfo:     "",
 		}
@@ -77,7 +77,6 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerCreateInvalidAlreadyExists
 			Developer:    "",
 			Version:      "v0.0.1",
 			SupportEmail: "test@email.com",
-			CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 			Enabled:      false,
 		}
 		_, err := srv.CreateCookbook(wctx, cookbook)
@@ -94,6 +93,7 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerCreateInvalidAlreadyExists
 			Entries:       types.EntriesList{},
 			Outputs:       nil,
 			BlockInterval: 0,
+			CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 			Enabled:       false,
 			ExtraInfo:     "",
 		}
@@ -124,7 +124,6 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerCreateInvalidCookbookNotOw
 			Developer:    "",
 			Version:      "v0.0.1",
 			SupportEmail: "test@email.com",
-			CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 			Enabled:      false,
 		}
 		_, err := srv.CreateCookbook(wctx, cookbook)
@@ -141,6 +140,7 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerCreateInvalidCookbookNotOw
 			Entries:       types.EntriesList{},
 			Outputs:       nil,
 			BlockInterval: 0,
+			CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 			Enabled:       false,
 			ExtraInfo:     "",
 		}
@@ -200,7 +200,6 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerUpdate() {
 		Developer:    "",
 		Version:      "v0.0.1",
 		SupportEmail: "test@email.com",
-		CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 		Enabled:      false,
 	}
 	_, err := srv.CreateCookbook(wctx, cookbook)
@@ -217,6 +216,7 @@ func (suite *IntegrationTestSuite) TestRecipeMsgServerUpdate() {
 		Entries:       types.EntriesList{},
 		Outputs:       nil,
 		BlockInterval: 0,
+		CostPerBlock: sdk.Coin{Denom: "test", Amount: sdk.ZeroInt()},
 		Enabled:       false,
 		ExtraInfo:     "",
 	}